requirements.txt
# --- Core UI & Framework ---
streamlit                  
pandas                    

# --- AI & Computer Vision ---
ultralytics               
torch                      
torchvision                
opencv-python            
numpy                      
Pillow                     

# --- Cloud & Infrastructure ---
boto3                      
python-dotenv              
PyYAML                     
scikit-learn 
seaborn

scikit-learn 
plotly